// src/models/models.ts
export interface NutritionalInfo {
    energy: number; carbohydrates: number; sugars: number; fiber: number;
    starch: number; proteins: number; total_fat: number; saturated_fat: number;
    trans_fat: number; monounsaturated_fat: number; polyunsaturated_fat: number;
    cholesterol: number; water: number; vitamin_a: number; vitamin_b1: number;
    vitamin_b2: number; vitamin_b3: number; vitamin_b5: number; vitamin_b6: number;
    vitamin_b7: number; vitamin_b9: number; vitamin_b12: number; vitamin_c: number;
    vitamin_d: number; vitamin_e: number; vitamin_k: number; calcium: number;
    iron: number; magnesium: number; phosphorus: number; potassium: number;
    zinc: number; sodium: number; copper: number; manganese: number;
    selenium: number; iodine: number; chromium: number; molybdenum: number;
    silicon: number; vanadium: number; omega_3: number; omega_6: number;
}
// ...
export interface HealthProfile {
    healthConditions: string[];
    allergies: string[];
    currentMedications: string[];
    dietaryRestrictions: string[];
    favoriteFoods?: string[];
    activityLevel: string;
    isActivityLevelDetailed: boolean;
    activityLevelDetail: string;
    bodyFatLevel: string; // 1. Magro(a) | 2. Atlético(a) | 3. Em Forma | 4. Acima do Peso;
    bodyFatPercentage: number;
    metabolism?: string;
    sex: string;
    height: number;
    weight: number;
    dateOfBirth: string;
}